refactor: add common/data/util
